tests were added in SearchActivityTest
