+++ app/build.gradle
@@ -55,6 +55,7 @@ dependencies {
     implementation 'com.google.code.gson:gson:2.10.1'
 
     implementation 'com.github.bumptech.glide:glide:4.15.1'
+    implementation 'androidx.test.espresso:espresso-contrib:3.5.1'
     kapt 'com.github.bumptech.glide:compiler:4.15.1'
 
     testImplementation 'junit:junit:4.13.2'
