+++ app/src/androidTest/java/bes/max/playlistmaker/AppTest.kt
@@ -14,7 +14,7 @@ import org.junit.Assert.*
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 @RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
+class AppTest {
     @Test
     fun useAppContext() {
         // Context of the app under test.
