+++ app/src/androidTest/java/bes/max/playlistmaker/SettingsActivityTest.kt
@@ -1,7 +1,6 @@
 package bes.max.playlistmaker
 
 import android.app.Activity
-import android.app.Instrumentation
 import android.content.Intent
 import android.net.Uri
 import androidx.lifecycle.Lifecycle
@@ -10,7 +9,6 @@ import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.Intents.intended
-import androidx.test.espresso.intent.Intents.intending
 import androidx.test.espresso.intent.matcher.IntentMatchers.*
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -18,7 +16,6 @@ import androidx.test.filters.LargeTest
 import bes.max.playlistmaker.presentation.settings.SettingsActivity
 import org.hamcrest.CoreMatchers.allOf
 import org.hamcrest.CoreMatchers.`is`
-import org.hamcrest.CoreMatchers.not
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers
 import org.junit.After
